refactor: 코드 가독성을 위해 서비스와 컨트롤러 분리
