+++ src/main/java/com/tave/tavewebsite/domain/member/controller/ManagerController.java
@@ -3,6 +3,7 @@ package com.tave.tavewebsite.domain.member.controller;
 import com.tave.tavewebsite.domain.member.dto.request.ResetPasswordReq;
 import com.tave.tavewebsite.domain.member.dto.request.ValidateEmailReq;
 import com.tave.tavewebsite.domain.member.dto.response.CheckNickNameResponseDto;
+import com.tave.tavewebsite.domain.member.service.AdminService;
 import com.tave.tavewebsite.domain.member.service.MemberService;
 import com.tave.tavewebsite.global.success.SuccessResponse;
 import lombok.RequiredArgsConstructor;
@@ -21,6 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class ManagerController {
 
     private final MemberService memberService;
+    private final AdminService adminService;
 
     @PostMapping("/normal/authenticate/email")
     public SuccessResponse sendEmail(@RequestBody ValidateEmailReq requestDto) {
@@ -39,7 +41,7 @@ public class ManagerController {
 
     @GetMapping("/normal/upgrade/{memberId}")
     public SuccessResponse updateAuthentication(@PathVariable("memberId") String memberId) {
-        memberService.updateAuthentication(memberId);
+        adminService.updateAuthentication(memberId);
 
         return new SuccessResponse("update Success.");
     }
